Presentation Models
RequestShelterModel
We are receiving data from the front end in the form of a request to shelter an animal.
AnimalType, AnimalSize, AnimalName
ResponseShelterModel
We are sending a response to the request from the front end to the front end.
Is the animal accepted and a message to go along with it.