Long Statement
Did too much without iterative progress. Long statement:

A KennelSize enum was created so that it could be used in the ShelterKennelBO which was also created. This was developed to be used to capture data from the database relevant to sheltering animals.  (There is need to possible not have AnimalName and AnimalSize in this, and therefore removed.

KennelCapacityBO was also created to be used in the KennelCapacityService which allows for capturing relevant stats of the kennels like capacity, size , and which are vacant. This adds increased functionality for later use in future development and may decrease turn over time when a new w animal  is added to the shelter.
Originally the KennelCapacityBO did have logic to find the stats on Kennel Counts, however as a need to separate concerns and testing of JUST THE SERVICE LAYER, I have moved them to functions in the service layer. Although all the properties of kennelCapacity are assigned in one Method, I may break this up to account for better testing.

ShelterService which returns the ShelterResponseModel will allow for responding to any shelter requests and either assign an animal a kennel, or will say that no kennels are available. In additon, if size is equal to 0 then no kennel is provided.
Need to decide on what other cases may occur for this service.
Can I improve this service by not pulling data and creating objects by checking for no vacancy?