Reorganize Animal Service is complete
Interesting exercise. The organization of animals requires that the smallest size animals be removed from the largest size kennels to the smallest. Then the small animals from the medium kennels to the smallest if available. If not, then use the medium kennels.

ShelterAnimal call controller created
GetKennels call controller created